pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
markdown = "^3.6"
pdfkit = "^1.0.0"
markdown2pdf = "^0.1.4"
weasyprint = "^61.2"
markdown2 = "^2.4.13"
reportlab = "^4.1.0"
pypdf = "^4.1.0"
fpdf = "^1.7.2"
python-dotenv = "^1.0.1"
google-generativeai = "^0.4.1"
google-cloud-aiplatform = "^1.44.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
